refactor auth tests to own file
