+++ api/server/server_test.go
@@ -4,7 +4,6 @@ import (
 	"bytes"
 	"context"
 	"encoding/json"
-	"fmt"
 	"io"
 	"io/ioutil"
 	"net/http"
@@ -19,8 +18,6 @@ import (
 	"github.com/iron-io/functions/api/runner"
 	"github.com/iron-io/functions/api/runner/task"
 	"github.com/iron-io/functions/api/server/internal/routecache"
-	"github.com/iron-io/functions/common"
-	"github.com/spf13/viper"
 )
 
 var tmpBolt = "/tmp/func_test_bolt.db"
@@ -35,52 +32,23 @@ type Suite []struct {
 }
 
 var testSuite = Suite{
-		{"create my app", "POST", "/v1/apps", `{ "app": { "name": "myapp" } }`, http.StatusOK, 0},
-		{"list apps", "GET", "/v1/apps", ``, http.StatusOK, 0},
-		{"get app", "GET", "/v1/apps/myapp", ``, http.StatusOK, 0},
-		{"add myroute", "POST", "/v1/apps/myapp/routes", `{ "route": { "name": "myroute", "path": "/myroute", "image": "iron/hello" } }`, http.StatusOK, 1},
-		{"add myroute2", "POST", "/v1/apps/myapp/routes", `{ "route": { "name": "myroute2", "path": "/myroute2", "image": "iron/error" } }`, http.StatusOK, 2},
-		{"get myroute", "GET", "/v1/apps/myapp/routes/myroute", ``, http.StatusOK, 2},
-		{"get myroute2", "GET", "/v1/apps/myapp/routes/myroute2", ``, http.StatusOK, 2},
-		{"get all routes", "GET", "/v1/apps/myapp/routes", ``, http.StatusOK, 2},
-		{"execute myroute", "POST", "/r/myapp/myroute", `{ "name": "Teste" }`, http.StatusOK, 2},
-		{"execute myroute2", "POST", "/r/myapp/myroute2", `{ "name": "Teste" }`, http.StatusInternalServerError, 2},
-		{"delete myroute", "DELETE", "/v1/apps/myapp/routes/myroute", ``, http.StatusOK, 1},
-		{"delete app (fail)", "DELETE", "/v1/apps/myapp", ``, http.StatusBadRequest, 1},
-		{"delete myroute2", "DELETE", "/v1/apps/myapp/routes/myroute2", ``, http.StatusOK, 0},
-		{"delete app (success)", "DELETE", "/v1/apps/myapp", ``, http.StatusOK, 0},
-		{"get deleted app", "GET", "/v1/apps/myapp", ``, http.StatusNotFound, 0},
-		{"get deleteds route on deleted app", "GET", "/v1/apps/myapp/routes/myroute", ``, http.StatusNotFound, 0},
-	}
-
-var UnAuthtestSuite = []struct {
-		name              string
-		method            string
-		path              string
-		body              string
-		expectedCode      int
-		expectedCacheSize int
-	}{
-		{"create my app", "POST", "/v1/apps", `{ "app": { "name": "myapp" } }`, http.StatusUnauthorized, 0},
-		{"list apps", "GET", "/v1/apps", ``, http.StatusUnauthorized, 0},
-		{"get app", "GET", "/v1/apps/myapp", ``, http.StatusUnauthorized, 0},
-		{"add myroute", "POST", "/v1/apps/myapp/routes", `{ "route": { "name": "myroute", "path": "/myroute", "image": "iron/hello" } }`, http.StatusUnauthorized, 0},
-		{"add myroute2", "POST", "/v1/apps/myapp/routes", `{ "route": { "name": "myroute2", "path": "/myroute2", "image": "iron/error" } }`, http.StatusUnauthorized, 0},
-		{"get myroute", "GET", "/v1/apps/myapp/routes/myroute", ``, http.StatusUnauthorized, 0},
-		{"get myroute2", "GET", "/v1/apps/myapp/routes/myroute2", ``, http.StatusUnauthorized, 0},
-		{"get all routes", "GET", "/v1/apps/myapp/routes", ``, http.StatusUnauthorized, 0},
-		// These two are currently returning 404 because they dont get created : temporarily using StatusNotFound
-		//		{"execute myroute", "POST", "/r/myapp/myroute", `{ "name": "Teste" }`, http.StatusUnauthorized, 0},
-		//		{"execute myroute2", "POST", "/r/myapp/myroute2", `{ "name": "Teste" }`, http.StatusUnauthorized, 0},
-		{"execute myroute", "POST", "/r/myapp/myroute", `{ "name": "Teste" }`, http.StatusNotFound, 0},
-		{"execute myroute2", "POST", "/r/myapp/myroute2", `{ "name": "Teste" }`, http.StatusNotFound, 0},
-		{"delete myroute", "DELETE", "/v1/apps/myapp/routes/myroute", ``, http.StatusUnauthorized, 0},
-		{"delete app (fail)", "DELETE", "/v1/apps/myapp", ``, http.StatusUnauthorized, 0},
-		{"delete myroute2", "DELETE", "/v1/apps/myapp/routes/myroute2", ``, http.StatusUnauthorized, 0},
-		{"delete app (success)", "DELETE", "/v1/apps/myapp", ``, http.StatusUnauthorized, 0},
-		{"get deleted app", "GET", "/v1/apps/myapp", ``, http.StatusUnauthorized, 0},
-		{"get deleteds route on deleted app", "GET", "/v1/apps/myapp/routes/myroute", ``, http.StatusUnauthorized, 0},
-	}
+	{"create my app", "POST", "/v1/apps", `{ "app": { "name": "myapp" } }`, http.StatusOK, 0},
+	{"list apps", "GET", "/v1/apps", ``, http.StatusOK, 0},
+	{"get app", "GET", "/v1/apps/myapp", ``, http.StatusOK, 0},
+	{"add myroute", "POST", "/v1/apps/myapp/routes", `{ "route": { "name": "myroute", "path": "/myroute", "image": "iron/hello" } }`, http.StatusOK, 1},
+	{"add myroute2", "POST", "/v1/apps/myapp/routes", `{ "route": { "name": "myroute2", "path": "/myroute2", "image": "iron/error" } }`, http.StatusOK, 2},
+	{"get myroute", "GET", "/v1/apps/myapp/routes/myroute", ``, http.StatusOK, 2},
+	{"get myroute2", "GET", "/v1/apps/myapp/routes/myroute2", ``, http.StatusOK, 2},
+	{"get all routes", "GET", "/v1/apps/myapp/routes", ``, http.StatusOK, 2},
+	{"execute myroute", "POST", "/r/myapp/myroute", `{ "name": "Teste" }`, http.StatusOK, 2},
+	{"execute myroute2", "POST", "/r/myapp/myroute2", `{ "name": "Teste" }`, http.StatusInternalServerError, 2},
+	{"delete myroute", "DELETE", "/v1/apps/myapp/routes/myroute", ``, http.StatusOK, 1},
+	{"delete app (fail)", "DELETE", "/v1/apps/myapp", ``, http.StatusBadRequest, 1},
+	{"delete myroute2", "DELETE", "/v1/apps/myapp/routes/myroute2", ``, http.StatusOK, 0},
+	{"delete app (success)", "DELETE", "/v1/apps/myapp", ``, http.StatusOK, 0},
+	{"get deleted app", "GET", "/v1/apps/myapp", ``, http.StatusNotFound, 0},
+	{"get deleteds route on deleted app", "GET", "/v1/apps/myapp/routes/myroute", ``, http.StatusNotFound, 0},
+}
 
 func testServer(ds models.Datastore, mq models.MessageQueue, rnr *runner.Runner, tasks chan task.Request) *Server {
 	ctx := context.Background()
@@ -115,32 +83,6 @@ func routerRequest(t *testing.T, router *gin.Engine, method, path string, body i
 	router.ServeHTTP(rec, req)
 
 	return req, rec
-}
-
-func routerRequestWithAuth(t *testing.T, router *gin.Engine, method, path string, body io.Reader, setAuth func(*http.Request)) (*http.Request, *httptest.ResponseRecorder) {
-	req, err := http.NewRequest(method, "http://127.0.0.1:8080"+path, body)
-	setAuth(req)
-	if err != nil {
-		t.Fatalf("Test: Could not create %s request to %s: %v", method, path, err)
-	}
-
-	rec := httptest.NewRecorder()
-	router.ServeHTTP(rec, req)
-
-	return req, rec
-}
-
-func setJwtAuth(req *http.Request) {
-	if jwtAuthKey := viper.GetString("jwt_auth_key"); jwtAuthKey != "" {
-		jwtToken, err := common.GetJwt(jwtAuthKey, 60*60)
-		if err == nil {
-			req.Header.Set("Authorization", fmt.Sprintf("Bearer %v", jwtToken))
-		}
-	}
-}
-
-func setBrokenJwtAuth(req *http.Request) {
-	req.Header.Set("Authorization", fmt.Sprintf("Bearer %v", "broken token"))
 }
 
 func newRouterRequest(t *testing.T, method, path string, body io.Reader) (*http.Request, *httptest.ResponseRecorder) {
@@ -182,16 +124,6 @@ func prepareBolt(t *testing.T) (models.Datastore, func()) {
 
 func TestFullStackWithNoAuth(t *testing.T) {
 	testFullStack(t, setJwtAuth, testSuite)
-}
-
-func TestFullStackWithAuth(t *testing.T) {
-	viper.Set("jwt_auth_key", "test")
-	testFullStack(t, setJwtAuth, testSuite)
-}
-
-func TestFullStackWithBrokenAuth(t *testing.T) {
-	viper.Set("jwt_auth_key", "test")
-	testFullStack(t, setBrokenJwtAuth, UnAuthtestSuite)
 }
 
 func testFullStack(t *testing.T, authFn func(*http.Request), suite Suite) {
